A HTTP server for a Host

Normally there is no need to create a new HostHttpServer, instead use the start method of the host instance.

HostHttpServer

Format

R6Class.

new()

host

The host to be served

port

The port to listen on

start()

Create a new HostHttpServer

Start the server

stop()

Stop the server

handle()

Handle a HTTP request

route()

Route a HTTP request

options()

Handle OPTIONS request Necessary for preflighted CORS requests (https://developer.mozilla.org/en-US/docs/Web/HTTP/Access_control_CORS#Preflighted_requests)

home()

Handle a request to `home`

static()

Handle a request for a static file

post()

Handle a POST request

Handle a GET request

put()

Handle a PUT request