A R context

An execution context for R code

RContext

Format

R6Class.

Details

In Stencila, a "context" is the thing that executes code for a particular programming language. This is the context for R. It implements the Context API so that it can talk to other parts of the platform, including contexts for other languages, Documents, and Sheets.

new()

Create a new RContext

Currently the parameter closed defaults to FALSE so that you can use library(somepackage) to make a package available in subsequent calls to runCode or callCode. In the future, it would be good to have a better machanism for that.

local

Context can not assign to the global environment. Default TRUE

closed

Context can not read from the global environment. Default FALSE

analyseCode()

Analyse R code and return the names of inputs, outputs and the implicitly returned vaue expression

code

R code to be analysed

exprOnly

Ensure that the code is a simple expression?

executeCode()

Run R code within the context's scope

code

R code to be executed

inputs

A list with a data pack for each input

exprOnly

Ensure that the code is a simple expression?

callCode()

Run R code within a local function scope

code

R code to be executed

inputs

A list with a data pack for each input

isolated

Is the call isolated from the context's global environment

Examples

context <- RContext$new()

# Assign a variable within the context
context$runCode('my_var <- 42')

#> Error in eval(expr, envir, enclos): attempt to apply non-function

# Get the variable as an output value
context$runCode('my_var')

#> Error in eval(expr, envir, enclos): attempt to apply non-function

# The variable is NOT available in `callCode`
context$callCode('my_var')$errors[[1]]$message

#> NULL

# Intead, you can pass input values
context$callCode('x * y', list(x=pack(6), y=pack(7)))

#> $value
#> NULL
#> 
#> $messages
#> $messages[[1]]
#> $messages[[1]]$line
#> [1] 1
#> 
#> $messages[[1]]$column
#> [1] 0
#> 
#> $messages[[1]]$type
#> [1] "error"
#> 
#> $messages[[1]]$message
#> [1] "object 'x' not found"
#> 
#> 
#> 

# Returned output value can include plots
context$callCode('plot(1,1)')$output

#> NULL