A SQLite context

A SQLite context

SqliteContext

Format

An object of class R6ClassGenerator of length 25.

new()

Create a new SqliteContext

dir

Working directory. Default NULL

compile()

Analyse SQL code and return the names of inputs, outputs and the returned value

code

SQL code to be analysed

exprOnly

Ensure that the code is a select expression?

execute()

Run SQL code

code

SQL code to be executed

inputs

A list with a value pack for each input

exprOnly

Ensure that the code is a simple expression?